pyproject.toml
[tool.poetry]
name = "django-lms-project"
version = "0.1.0"
description = ""
authors = ["Swetlana Mirzoeva <levka008@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.16.1"
pillow = "^12.1.0"
python-dotenv = "^1.2.1"
django-filter = "^25.2"
djangorestframework-simplejwt = "^5.5.1"


[tool.poetry.group.lint.dependencies]
black = "^26.1.0"
isort = "^7.0.0"
mypy = "^1.19.1"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
